refactor(graph-model): pre-extract graph domain model + proto conversions to foundation
Move src/graph/model.rs (273 lines, 20+ types: Node, Edge, PropertyValue,
EmbeddingVersion, TraversalRequest/Response, etc.) + src/graph/proto_convert.rs
(444 lines, ~32 From<proto> <-> model boundary conversions) to a NEW foundation
crate `proximadb-graph-model`. Both files are foundation-pure (HashMap + serde +
proto deps only).

Root files replaced with `pub use` re-exports — all callers unchanged.

Coherence fix: proto_convert.rs's `impl From<pb::Node> for m::Node` conversions
caused E0117 orphan-rule errors when model types moved (root can't impl for
external types). Moving proto_convert.rs to the crate resolves this — graph-model
depends on proximadb-proto (foundation) + defines the model types, so the From
impls are in the correct crate.

Resolves 2 of observability's 8 outbound cross-module deps (Edge + Node).
Combined with TenantStorageUsage (this branch's first commit), 3 of 8 resolved.
The graph-model crate is a shared foundation dependency (graph engine +
observability + all graph consumers).
