+++ crates/foundation/proximadb-graph-model/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "proximadb-graph-model"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Foundation graph domain model + proto conversions: Node, Edge, PropertyValue, TraversalRequest/Response, and the From<proto> boundary conversions"
+
+[lib]
+name = "proximadb_graph_model"
+path = "src/lib.rs"
+
+[dependencies]
+serde.workspace = true
+proximadb-proto.workspace = true
+
+[lints]
+workspace = true
