+++ crates/foundation/proximadb-graph-model/src/lib.rs
@@ -1,0 +1,4 @@
+//! Foundation graph domain model + proto conversions (moved from src/graph/).
+pub mod model;
+pub mod proto_convert;
+pub use model::*;
